Revert unwanted change

+++ shared/src/commonMain/kotlin/com/hicham/wcstoreapp/models/ProductPrices.kt
@@ -1,7 +1,9 @@
 package com.hicham.wcstoreapp.models
 
+import java.math.BigDecimal
+
 data class ProductPrices(
-    val price: Double,
-    val regularPrice: Double,
-    val salePrice: Double
+    val price: BigDecimal,
+    val regularPrice: BigDecimal,
+    val salePrice: BigDecimal
 )
